requirements.txt
langchain
langchain-community
langchain-openai
chromadb
streamlit
pandas
numpy
db-sqlite3
requests
python-dotenv
PyPDF2
beautifulsoup4
lxml
openai
faiss-cpu
tiktoken
click
tabulate
openpyxl
python-multipart